requirements.txt
imageio
scipy
argparse
Pillow
sacred
easydict
pymongo
pyyaml
matplotlib
opencv-python